fix(backend): address PR #2748 fixup round 1 review findings
Extract a shared RepositoryOrigin() helper to stop the executor and
lifecycle assembly sites from duplicating repository-origin
construction, move the test-only resolveEnvironmentSources helper
into task_environment_test.go now that it has no production caller,
assert the full AC-23 log field set in the override-precedence test,
and advance the spec's status to building with an INDEX.md entry now
that the feature is in active PR review.
