+++ docs/specs/executor-profile-env-precedence/spec.md
@@ -1,5 +1,5 @@
 ---
-status: draft
+status: building
 created: 2026-08-17
 owner: tbd
 ---
